chore: split test in separate file
