+++ src/tests/libecalc/presentation/yaml/yaml_types/models/conftest.py
@@ -1,35 +1,23 @@
-from io import StringIO
-from pathlib import Path
 from typing import Dict, Generic, List, Optional, TypeVar, Union, cast
 
-import pytest
 from models_test_setup import (
     ChartSingleSpeed,
     ChartVariableSpeed,
-    CompressorTrainBase,
     Curve,
     FluidModel,
-    Models,
     SimplifiedVariableSpeedTrainKnownStagesModel,
     SimplifiedVariableSpeedTrainUnknownStagesModel,
     SingleSpeedTrainModel,
     Stage,
-    Stages,
     VariableSpeedTrainModel,
 )
 from polyfactory import Require
 from polyfactory.decorators import post_generated
 from polyfactory.factories.pydantic_factory import ModelFactory
 
-from ecalc_cli.infrastructure.file_resource_service import FileResourceService
-from libecalc.common.time_utils import Frequency
 from libecalc.dto import GenericChartFromDesignPoint, GenericChartFromInput
 from libecalc.presentation.yaml.configuration_service import ConfigurationService
-from libecalc.presentation.yaml.model import YamlModel
-from libecalc.presentation.yaml.validation_errors import ValidationError
 from libecalc.presentation.yaml.yaml_entities import ResourceStream
-from libecalc.presentation.yaml.yaml_keywords import EcalcYamlKeywords
-from libecalc.presentation.yaml.yaml_models.pyyaml_yaml_model import PyYamlYamlModel
 from libecalc.presentation.yaml.yaml_models.yaml_model import ReaderType, YamlConfiguration, YamlValidator
 
 
@@ -182,61 +170,6 @@ class CurveFactory(ModelFactory[Curve]):
             efficiency_sample = cls.__random__.uniform(a=_min_efficiency, b=_max_efficiency)
             efficiency.append(efficiency_sample)
         return efficiency
-
-
-def test_single_speed_compressor_train() -> None:
-    curve_instance = CurveFactory.build(
-        _min_speed=3000,
-        _max_speed=6500,
-        _min_rate=3500,
-        _max_rate=7500,
-        _min_head=5000,
-        _max_head=9500,
-        _min_efficiency=0.7,
-        _max_efficiency=0.8,
-    )
-    chart_instance = SingleSpeedChartFactory.build(name="single_speed_chart", curve=curve_instance)
-    stage_instance = StageFactory.build(
-        compressor_chart=chart_instance.name, control_margin=0, control_margin_unit="FRACTION"
-    )
-    fluid_model_instance = FluidModelFactory.build(
-        name="fluid_model", fluid_model_type="PREDEFINED", gas_type="DRY", eos_model="SRK"
-    )
-    compressor_instance = SingleSpeedTrainFactory.build(
-        name="single_speed_compressor_train",
-        compressor_chart=chart_instance.name,
-        compressor_train=Stages(stages=[stage_instance]),
-        fluid_model=fluid_model_instance.name,
-    )
-    models = Models(
-        models=[
-            fluid_model_instance,
-            chart_instance,
-            compressor_instance,
-        ]
-    )
-    yaml_dict = models.model_dump(by_alias=True)
-    cleaned_data = remove_null_none_empty(yaml_dict)
-    yaml_string = PyYamlYamlModel.dump_yaml(yaml_dict=cleaned_data)
-
-    configuration_service = OverridableStreamConfigurationService(
-        stream=ResourceStream(name="", stream=StringIO(yaml_string))
-    )
-    resource_service = FileResourceService(working_directory=Path(""))
-
-    model = YamlModel(
-        configuration_service=configuration_service,
-        resource_service=resource_service,
-        output_frequency=Frequency.YEAR,
-    )
-
-    with pytest.raises(ValidationError) as exc_info:
-        model.validate_for_run()
-
-    assert f"{EcalcYamlKeywords.fuel_types}\n" f"\tMessage: This keyword is missing, it is required" in str(
-        exc_info.value
-    )
-    assert isinstance(compressor_instance, CompressorTrainBase)
 
 
 yaml_fluid_model = """
